requirements.txt
Django==5.1.4
psycopg2-binary==2.9.10
djangorestframework==3.15.2
pandas==2.2.3
openpyxl==3.1.5
numpy==2.2.6
whitenoise==6.8.2
